refactor: code review - fix bugs, security issues, and clippy warnings
Code review findings and fixes:

Bug fixes:
- quality_management: calculate_deviation() had identical if/else branches
  making the spec-limit check dead code. Now outputs 'OUT_OF_SPEC' indicator
  when outside specification limits.
- collections: unwrap() after is_some() replaced with if-let pattern to
  prevent potential panics.

Security fixes:
- approval_delegation handlers (cancel/activate/delete): Authorization checks
  used 'if let Ok(Some(rule))' which silently bypassed auth on DB errors.
  Now properly propagates errors and returns 404 for missing rules.

Clippy fixes (776 -> 469 warnings in atlas-core, 5 -> 1 in atlas-gateway):
- Removed 196 redundant reference/deref patterns
- Replaced 104 redundant closures with method references
- Fixed 15 manual !RangeInclusive::contains implementations
- Fixed 19 map().flatten() -> and_then() on Option
- Fixed 19 map_or -> is_none_or
- Fixed 3 unnecessary type casts (as i32 -> i32)
- Fixed wildcard pattern in descriptive_flexfield
- Fixed match-like-expr -> matches! macro in product_information
- Removed identical if/else in cross_validation handler
- Fixed 2 filter_map -> map in purchase_requisition handler
- Removed dead #[allow(dead_code)] on 19 used constants in quality_management

Remaining warnings are design-level (too_many_arguments, very_complex_type)
which require larger refactoring with builder/request structs.

+++ crates/atlas-core/src/subscription/engine.rs
@@ -1039,7 +1039,7 @@ impl SubscriptionEngine {
             let billing_date = self.calculate_billing_date(period_start, sub.billing_day_of_month);
 
             // Proration for last period if it's shorter than a full period
-            let full_period_days = months_per as i32 * 30;
+            let full_period_days = months_per * 30;
             let actual_days = (period_end - period_start).num_days() + 1;
             let proration_factor = if actual_days < full_period_days as i64 {
                 actual_days as f64 / full_period_days as f64
